requirements.txt
streamlit
pandas
scikit-learn
google-generativeai
fpdf2
numpy
plotly
python-dotenv
